# docs/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/v1/projects/{projectId}/game-modes": {
            "post": {
                "operationId": "GameModesController_create",
                "parameters": [
                    {
                        "name": "projectId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGameModeDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "sessionToken": []
                    }
                ],
                "summary": "Create a game mode for a project.",
                "tags": ["Game Modes"]
            },
            "get": {
                "operationId": "GameModesController_findAll",
                "parameters": [
                    {
                        "name": "projectId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "sessionToken": []
                    }
                ],
                "summary": "List game modes for a project.",
                "tags": ["Game Modes"]
            }
        },
...
        "/v1/projects/{projectId}/disputes": {
            "get": {
                "operationId": "DisputesController_findAll",
                "parameters": [
                    {
                        "name": "projectId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "description": "Filter by dispute status (OPEN, RESOLVED, REJECTED).",
                        "schema": {
                            "type": "string",
                            "enum": ["OPEN", "RESOLVED", "REJECTED"]
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "Maximum number of disputes to return.",
                        "schema": {
                            "minimum": 1,
                            "maximum": 100,
                            "default": 50,
                            "type": "number"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "Number of records to skip for pagination.",
                        "schema": {
                            "minimum": 0,
                            "default": 0,
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "sessionToken": []
                    }
                ],
                "summary": "List match disputes for a project with optional status filter.",
                "tags": ["Disputes"]
            }
        },
        "/v1/projects/{projectId}/disputes/{disputeId}": {
            "get": {
                "operationId": "DisputesController_findOne",
                "parameters": [
                    {
                        "name": "projectId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "disputeId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "sessionToken": []
                    }
                ],
                "summary": "Get detailed dispute information including match slots and evidence.",
                "tags": ["Disputes"]
            }
        },
        "/v1/projects/{projectId}/disputes/{disputeId}/resolve": {
            "post": {
                "operationId": "DisputesController_resolve",
                "parameters": [
                    {
                        "name": "projectId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "disputeId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResolveDisputeDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "sessionToken": []
                    }
                ],
                "summary": "Resolve a dispute with optional winner override and Elo rating reconciliation.",
                "tags": ["Disputes"]
            }
        },
        "/v1/projects/{projectId}/disputes/{disputeId}/reject": {
            "post": {
                "operationId": "DisputesController_reject",
                "parameters": [
                    {
                        "name": "projectId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "disputeId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RejectDisputeDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "sessionToken": []
                    }
                ],
                "summary": "Reject a dispute and restore match status.",
                "tags": ["Disputes"]
            }
        },
...
    },
    "info": {
        "title": "Matching Man API",
        "description": "Multi-tenant matchmaking platform API.\n\nGame servers authenticate with a project API key (`projectApiKey`) issued via `POST /v1/projects/:projectId/api-keys`.\nThe admin dashboard authenticates with a per-user session token from `/v1/auth/login` or `/v1/auth/register`, or the break-glass `DASHBOARD_ADMIN_TOKEN` (`sessionToken`).",
        "version": "1",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "projectApiKey": {
                "scheme": "bearer",
                "bearerFormat": "token",
                "type": "http",
                "description": "Project API key, e.g. `Authorization: Bearer <project_api_key>`."
            },
            "sessionToken": {
                "scheme": "bearer",
                "bearerFormat": "token",
                "type": "http",
                "description": "Dashboard session token or admin token, e.g. `Authorization: Bearer <session_token>`."
            }
        },
        "schemas": {
...
            "CreateGameModeDto": {
                "type": "object",
                "properties": {
                    "key": {
                        "type": "string",
                        "example": "ranked-5v5",
                        "maxLength": 120
                    },
                    "name": {
                        "type": "string",
                        "example": "Ranked 5v5",
                        "maxLength": 120
                    },
                    "matchStructure": {
                        "type": "string",
                        "enum": ["VERSUS", "FFA"]
                    },
                    "requiredSlots": {
                        "type": "number",
                        "minimum": 2,
                        "maximum": 64,
                        "description": "Total player slots across the match."
                    },
                    "groupCount": {
                        "type": "number",
                        "minimum": 2,
                        "maximum": 64,
                        "description": "Number of opposing groups (e.g. 2 for teams A/B)."
                    },
                    "teamSizeMin": {
                        "type": "number",
                        "minimum": 1,
                        "maximum": 64
                    },
                    "teamSizeMax": {
                        "type": "number",
                        "minimum": 1,
                        "maximum": 64
                    },
                    "ratingMode": {
                        "type": "string",
                        "enum": ["INTERNAL_ELO", "EXTERNAL_RATING", "DISABLED"]
                    },
                    "initialRatingWindow": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Initial rating band width for candidate matching."
                    },
                    "windowExpandIntervalSeconds": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Seconds between each rating window expansion."
                    },
                    "windowExpandStep": {
                        "type": "number",
                        "minimum": 1
                    },
                    "enableReadyCheck": {
                        "type": "boolean",
                        "default": false,
                        "description": "Whether matches require an accept/decline ready check handshake."
                    },
                    "readyCheckTimeoutSeconds": {
                        "type": "number",
                        "default": 20,
                        "minimum": 5,
                        "maximum": 300,
                        "description": "Seconds players have to accept a match."
                    }
                },
                "required": [
                    "key",
                    "name",
                    "matchStructure",
                    "requiredSlots",
                    "groupCount",
                    "teamSizeMin",
                    "teamSizeMax"
                ]
            },
...
            "ResolveDisputeDto": {
                "type": "object",
                "properties": {
                    "overrideWinnerGroupIndex": {
                        "type": "object",
                        "minimum": 1,
                        "example": 2,
                        "description": "The winning group index override (1-based), or omit/null to declare a draw/void."
                    },
                    "resolutionNotes": {
                        "type": "string",
                        "example": "Reviewed match replay clip. Player in team 1 was using prohibited client modifications.",
                        "maxLength": 1000,
                        "description": "Mandatory operator resolution notes explaining the rationale for the resolution."
                    }
                },
                "required": ["resolutionNotes"]
            },
            "RejectDisputeDto": {
                "type": "object",
                "properties": {
                    "resolutionNotes": {
                        "type": "string",
                        "example": "Insufficient evidence provided. Replay shows legitimate game behavior.",
                        "maxLength": 1000,
                        "description": "Mandatory operator resolution notes explaining the rationale for rejecting the dispute."
                    }
                },
                "required": ["resolutionNotes"]
            }
        }
    }
}